Refactor semester management to use dedicated form screen
